Move method to utils.
