+++ packages/devtools/lib/src/timeline/cpu_profile_protocol.dart
@@ -1,8 +1,6 @@
 // Copyright 2019 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
-import 'package:meta/meta.dart';
-
 import '../utils.dart';
 import 'cpu_profile_model.dart';
 
@@ -71,30 +69,5 @@ class CpuProfileProtocol {
       );
       cpuProfileData.stackFrames[leafId]?.exclusiveSampleCount++;
     }
-  }
-
-  /// Returns a simplified version of a StackFrame name.
-  ///
-  /// Given an input such as
-  /// "_WidgetsFlutterBinding&BindingBase&GestureBinding.handleBeginFrame", this
-  /// method will strip off all the Mixin names and return
-  /// "_WidgetsFlutterBinding.handleBeginFrame".
-  @visibleForTesting
-  String getSimpleStackFrameName(String name) {
-    final firstAmpersandIndex = name.indexOf('&');
-    final firstPeriodIndex = name.indexOf('.');
-
-    if (firstAmpersandIndex != -1 &&
-        firstPeriodIndex != -1 &&
-        firstAmpersandIndex < firstPeriodIndex &&
-        name.length > firstAmpersandIndex + 1) {
-      final nextCharCodeUnit = name[firstAmpersandIndex + 1].codeUnitAt(0);
-      if (isLetter(nextCharCodeUnit) || nextCharCodeUnit == '_'.codeUnitAt(0)) {
-        return name.substring(0, firstAmpersandIndex) +
-            name.substring(firstPeriodIndex);
-      }
-    }
-
-    return name;
   }
 }
